+++ backend/services/taskService.go
@@ -10,9 +10,9 @@ import (
 	"time"
 )
 
-func GetAllTasks() ([]models.Task, error) {
-	query := "SELECT * FROM tasks"
-	rows, err := database.DB.Query(query)
+func GetAllTasks(userId int64) ([]models.Task, error) {
+	query := "SELECT * FROM tasks WHERE user_id = $1"
+	rows, err := database.DB.Query(query, userId)
 	if err != nil {
 		fmt.Println("error fetching:", err)
 		return nil, err
@@ -46,9 +46,9 @@ func GetAllTasks() ([]models.Task, error) {
 	return tasks, nil
 }
 
-func GetAllProgressTasks() ([]models.Task, error) {
-	query := "SELECT * FROM tasks WHERE status = $1"
-	rows, err := database.DB.Query(query, models.StatusProgress)
+func GetAllProgressTasks(userId int64) ([]models.Task, error) {
+	query := "SELECT * FROM tasks WHERE status = $1 AND user_id = $2"
+	rows, err := database.DB.Query(query, models.StatusProgress, userId)
 	if err != nil {
 		fmt.Println("error fetching:", err)
 		return nil, err
@@ -82,15 +82,16 @@ func GetAllProgressTasks() ([]models.Task, error) {
 	return tasks, nil
 }
 
-func getTaskById(id int) (*models.Task, error) {
+func getTaskById(id int64, userId int64) (*models.Task, error) {
 
 	query := `
 		SELECT id, title, content, status, is_repeat, interval, notes, due_date, exec_at, created_at, updated_at, user_id
 		FROM tasks
 		WHERE id = $1
+		AND user_id = $2
 	`
 
-	row := database.DB.QueryRow(query, id)
+	row := database.DB.QueryRow(query, id, userId)
 
 	var task models.Task
 	err := row.Scan(
@@ -118,7 +119,7 @@ func getTaskById(id int) (*models.Task, error) {
 	return &task, nil
 }
 
-func CreateTask(task *models.Task) error {
+func CreateTask(task *models.Task, userId int64) error {
 	task.CreatedAt = time.Now()
 	task.UpdatedAt = time.Now()
 	task.Status = models.StatusProgress
@@ -142,7 +143,7 @@ func CreateTask(task *models.Task) error {
 		task.ExecAt,
 		task.CreatedAt,
 		task.UpdatedAt,
-		task.UserID,
+		userId,
 	).Scan(&task.ID)
 
 	if err != nil {
@@ -153,15 +154,16 @@ func CreateTask(task *models.Task) error {
 	return nil
 }
 
-func UpdateTask(task *models.Task) error {
+func UpdateTask(task *models.Task, userId int64) error {
 	task.UpdatedAt = time.Now()
 	query := `
 		UPDATE tasks
 		SET title = $1, content = $2, is_repeat = $3, interval = $4, due_date = $5,   updated_at = $6
 		WHERE id = $7
+		AND user_id = $8
 	`
 
-	_, err := database.DB.Exec(query, task.Title, task.Content, task.IsRepeat, task.Interval, task.DueDate, task.UpdatedAt, task.ID)
+	_, err := database.DB.Exec(query, task.Title, task.Content, task.IsRepeat, task.Interval, task.DueDate, task.UpdatedAt, task.ID, userId)
 	if err != nil {
 		fmt.Println("update error:", err)
 		return err
@@ -170,11 +172,11 @@ func UpdateTask(task *models.Task) error {
 	return nil
 }
 
-func DeleteTaskById(id int) error {
+func DeleteTaskById(id int64, userId int64) error {
 
-	query := `DELETE FROM tasks WHERE id = $1`
+	query := `DELETE FROM tasks WHERE id = $1 AND user_id = $2`
 
-	_, err := database.DB.Exec(query, id)
+	_, err := database.DB.Exec(query, id, userId)
 	if err != nil {
 		return fmt.Errorf("failed to delete task: %w", err)
 	}
@@ -182,7 +184,7 @@ func DeleteTaskById(id int) error {
 	return nil
 }
 
-func CompleteTaskById(id int, taskDto dto.EndTaskDTO) error {
+func CompleteTaskById(id int64, taskDto dto.EndTaskDTO, userId int64) error {
 
 	status := models.StatusDone
 	now := time.Now()
@@ -194,14 +196,15 @@ func CompleteTaskById(id int, taskDto dto.EndTaskDTO) error {
 			exec_at = $3,
 			updated_at = $4
 		WHERE id = $5
+		AND user_id = $6
 	`
-	_, err := database.DB.Exec(query, status, taskDto.Notes, now, now, id)
+	_, err := database.DB.Exec(query, status, taskDto.Notes, now, now, id, userId)
 	if err != nil {
 		fmt.Println("Abort error: ", err)
 		return err
 	}
 
-	task, _ := getTaskById(id)
+	task, _ := getTaskById(id, userId)
 	if task.IsRepeat {
 		var newTask models.Task
 		newTask.ParentID = task.ID
@@ -212,7 +215,7 @@ func CompleteTaskById(id int, taskDto dto.EndTaskDTO) error {
 		newTask.Notes = taskDto.Notes
 		newTask.DueDate = utils.AddDays(now, int(task.Interval))
 
-		err = CreateTask(&newTask)
+		err = CreateTask(&newTask, userId)
 		if err != nil {
 			return fmt.Errorf("Error creating new task after completion: %w", err)
 		}
@@ -221,7 +224,7 @@ func CompleteTaskById(id int, taskDto dto.EndTaskDTO) error {
 	return nil
 }
 
-func AbortTaskById(id int, taskDto dto.EndTaskDTO) error {
+func AbortTaskById(id int64, taskDto dto.EndTaskDTO, userId int64) error {
 
 	status := models.StatusAborted
 	now := time.Now()
@@ -233,14 +236,15 @@ func AbortTaskById(id int, taskDto dto.EndTaskDTO) error {
 			exec_at = $3,
 			updated_at = $4
 		WHERE id = $5
+		AND user_id = $6
 	`
-	_, err := database.DB.Exec(query, status, taskDto.Notes, now, now, id)
+	_, err := database.DB.Exec(query, status, taskDto.Notes, now, now, id, userId)
 	if err != nil {
 		fmt.Println("Abort error: ", err)
 		return err
 	}
 
-	task, _ := getTaskById(id)
+	task, _ := getTaskById(id, userId)
 	if task.IsRepeat {
 		var newTask models.Task
 		newTask.ParentID = task.ID
@@ -251,7 +255,7 @@ func AbortTaskById(id int, taskDto dto.EndTaskDTO) error {
 		newTask.Notes = taskDto.Notes
 		newTask.DueDate = utils.AddDays(now, int(task.Interval))
 
-		err = CreateTask(&newTask)
+		err = CreateTask(&newTask, userId)
 		if err != nil {
 			return fmt.Errorf("Error creating new task after abort: %w", err)
 		}
